Some docs and change of --lb meaning
rather than a branch length it is now the path length from
A to the closest node attached to the short internal edge.
This just means that the 4-taxon version can be produced by
pruning A1 and C1 from the tree (to result in a simulation
like you would see if you called the 4 taxon version with
the same --lb and --sb args).